+++ wemake_python_styleguide/constants.py
@@ -161,6 +161,9 @@ MAGIC_NUMBERS_WHITELIST: Final = frozenset((
 #: Maximum amount of ``# noqa`` comments per module.
 MAX_NOQA_COMMENTS: Final = 10
 
+#: Maximum amount of ``# pragma: no cover`` comments per module.
+MAX_NO_COVER_COMMENTS: Final = 5
+
 
 # Internal variables
 # They are not publicly documented since they are not used by the end user.